+++ pkg/annotations/utils.go
@@ -2,21 +2,15 @@ package annotations
 
 const annotationQueueName = "annotation-queue"
 const workerThreadCount = 1
-const WorkQueueRetryLimit = 3
+const workQueueRetryLimit = 3
 
-type info struct {
-	RKind string
-	RNs   string
-	RName string
-	//TODO:Hack as slice makes the struct unhasable
-	Patch *[]byte
+func getStatus(status bool) string {
+	if status {
+		return "Success"
+	}
+	return "Failure"
 }
 
-func newInfo(rkind, rns, rname string, patch []byte) info {
-	return info{
-		RKind: rkind,
-		RNs:   rns,
-		RName: rname,
-		Patch: &patch,
-	}
+func BuildKey(policyName string) string {
+	return "policies.kyverno.io." + policyName
 }
